fixing interfaces organization
