+++ src/core/domain/issues/interfaces/issues.interface.ts
@@ -29,33 +29,3 @@ export interface IIssue {
     createdAt: string;
     updatedAt: string;
 }
-
-export interface IIssueDetails {
-    title: string;
-    description: string;
-    age: string;
-    label: LabelType;
-    severity: SeverityLevel;
-    status: IssueStatus;
-    fileLink: {
-        label: string;
-        url: string;
-    };
-    prLinks: {
-        label: string;
-        url: string;
-    }[];
-    repositoryLink: {
-        label: string;
-        url: string;
-    };
-    currentCode: string;
-    language: string;
-    startLine: number;
-    endLine: number;
-    reactions: {
-        thumbsUp: number;
-        thumbsDown: number;
-    };
-    gitOrganizationName: string;
-}
